WEB-4654 update imports for consistency
